Moved SourceStringAnalyzerTest from src/test/resources/ to src/test/groovy so that IntelliJ can run all the unit tests without an error.
Please tell me if this change is ill-advised. It looks like the file was accidentally created in the wrong directory, but I might be missing the point of the file.
