+++ src/test/groovy/org/codenarc/analyzer/SourceStringAnalyzerTest.groovy
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2014 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.codenarc.analyzer
+
+import org.codenarc.results.Results
+import org.codenarc.rule.naming.ClassNameRule
+import org.codenarc.ruleset.ListRuleSet
+import org.codenarc.test.AbstractTestCase
+import org.junit.Test
+
+/**
+ * Test for SourceStringAnalyzer.
+ * @author Hamlet D'Arcy
+ */
+class SourceStringAnalyzerTest extends AbstractTestCase {
+
+    @Test
+    void testRunAgainstString() {
+        def source = '''
+            class badName { }
+        '''
+        def ruleSet = new ListRuleSet([new ClassNameRule()])
+        Results results = new StringSourceAnalyzer(source).analyze(ruleSet)
+        log(results.violations)
+        assert results.violations*.rule.name == ['ClassName']
+    }
+}
