--- src/test/resources/org/codenarc/analyzer/SourceStringAnalyzerTest.groovy
@@ -1,25 +1,0 @@
-package org.codenarc.analyzer
-
-import org.codenarc.results.Results
-import org.codenarc.rule.naming.ClassNameRule
-import org.codenarc.ruleset.ListRuleSet
-import org.codenarc.test.AbstractTestCase
-import org.junit.Test
-
-/**
- * Test for SourceStringAnalyzer.
- * @author Hamlet D'Arcy
- */
-class SourceStringAnalyzerTest extends AbstractTestCase {
-
-    @Test
-    void testRunAgainstString() {
-        def source = '''
-            class badName { }
-        '''
-        def ruleSet = new ListRuleSet([new ClassNameRule()])
-        Results results = new StringSourceAnalyzer(source).analyze(ruleSet)
-        log(results.violations)
-        assert results.violations*.rule.name == ['ClassName']
-    }
-}
